+++ R/to_sparse.R
@@ -1,3 +1,15 @@
+#' Convert to sparse
+#' 
+#' Convert a \link[base]{matrix} / \link[base]{data.frame} to a sparse matrix.
+#' If \code{X} is already a sparse matrix, it simply returns \code{X} directly.
+#' @param X A \link[base]{matrix} / \link[base]{data.frame}.
+#' @param verbose Print messages.
+#' @returns Sparse matrix.
+#' 
+#' @export
+#' @importFrom methods is as
+#' @examples 
+#' mat <- to_sparse(X = mtcars)
 to_sparse <- function(X,
                       verbose = TRUE) {
     messager("Converting obj to sparseMatrix.", v = verbose)