+++ tests/testthat/test-get_LD.R
@@ -11,8 +11,8 @@ test_that("get_LD works", {
         testthat::expect_gte(nrow(LD_list$LD), 40)
         testthat::expect_true(methods::is(LD_list$path, "character"))
         testthat::expect_true(file.exists(LD_list$path))
-    }
-
+    } 
+    
     #### 1000 Genomes: Phase 1 ####
     LD_1kgp1 <- echoLD::get_LD(
         locus_dir = locus_dir,
@@ -55,21 +55,47 @@ test_that("get_LD works", {
                              LD_reference = "UKB")
     run_tests(LD_ukb)
 
-    # # Local vcf file
-    # vcf_url <- file.path(
-    #     "ftp://ftp-trace.ncbi.nih.gov/1000genomes/ftp/release",
-    #     "20110521/",
-    #     paste0("ALL.chr",22,".phase1_release_v3.20101123",
-    #            ".snps_indels_svs.genotypes.vcf.gz"))
-    # out_paths <- downloadR::download_vcf(vcf_url = vcf_url,
-    # #                                      nThread = 1)
-    # LD_reference <- file.path(
-    #     "/var/folders/zq/h7mtybc533b1qzkys_ttgpth0000gn/T/RtmpQuMo8l",
-    #     "results/GWAS/Nalls23andMe_2019/BST1/LD/BST1.1KGphase3.vcf")
-    # LD_reference <- "var/folders/zq/h7mtybc533b1qzkys_ttgpth0000gn/T//RtmpQuMo8l/results/GWAS/Nalls23andMe_2019/BST1/LD/BST1.1KGphase3.vcf"
-    # file.exists(LD_reference)
-    # LD_local <- get_LD(locus_dir = locus_dir,
-    #                           query_dat= BST1,
-    #                            LD_reference = LD_reference,
-    #                            force_new_LD = TRUE)
+    #### Local vcf file ####
+    query_dat <-  echodata::BST1[seq(1, 50), ]
+    locus_dir <- echodata::locus_dir
+    locus_dir <- file.path(tempdir(), locus_dir)
+    LD_reference <- system.file("extdata", "BST1.1KGphase3.vcf.bgz",
+        package = "echodata")
+    LD_list_local <- get_LD(locus_dir = locus_dir,
+                            query_dat = query_dat,
+                            LD_reference = LD_reference)
+    run_tests(LD_list = LD_list_local)
+    
+    #### Remote vcf file ####  
+    LD_reference <- paste(
+        "https://github.com/RajLabMSSM/echodata/raw/main",
+        "inst/extdata/BST1.1KGphase3.vcf.bgz",sep="/"
+    )  
+    LD_list_remote <- get_LD(locus_dir = locus_dir,
+                             query_dat= query_dat,
+                             LD_reference = LD_reference,
+                             force_new_LD = TRUE)
+    run_tests(LD_list = LD_list_remote)
+    
+    
+    #### Local matrix file #### 
+    LD_reference <- tempfile(fileext = ".csv")
+    utils::write.csv(echodata::BST1_LD_matrix,
+                     file = LD_reference,
+                     row.names = TRUE)
+    LD_list_matrix_local <- get_LD_matrix(
+        locus_dir = locus_dir,
+        query_dat = query_dat,
+        LD_reference = LD_reference)
+    run_tests(LD_list = LD_list_matrix_local)
+    
+    
+    #### Remote matrix file #### 
+    LD_reference <- 
+        "https://github.com/RajLabMSSM/echodata/raw/main/data/BST1_LD_matrix.rda" 
+    LD_list_matrix_remote <- get_LD_matrix(
+        locus_dir = locus_dir,
+        query_dat = query_dat,
+        LD_reference = LD_reference)
+    run_tests(LD_list = LD_list_matrix_remote) 
 })